Move CreateExtraKey from local to remote
